requirements.txt
langchain-groq 
pandas 
chromadb 
sentence-transformers
fastapi
uvicorn
langchain
langchain-community
python-dotenv